Start testing DefaultViewportTransition
